move view utility functions to a new header and move some more forward declarations to ViewFwd
